requirements.txt
streamlit==1.27.0
pandas==2.1.0
numpy==1.25.2
matplotlib==3.8.0
seaborn==0.12.2
pillow==10.0.1
firebase-admin==6.2.0
python-dotenv==1.0.0
pyinstaller==6.0.0
